requirements/anaconda_rmac.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
absl-py=0.12.0=pypi_0
astor=0.8.1=pypi_0
blas=1.0=mkl
ca-certificates=2021.1.19=h06a4308_1
certifi=2020.6.20=pyhd3eb1b0_3
gast=0.4.0=pypi_0
google-pasta=0.2.0=pypi_0
grpcio=1.36.1=pypi_0
h5py=2.10.0=pypi_0
importlib-metadata=2.1.1=pypi_0
intel-openmp=2019.4=243
keras=2.2.4=pypi_0
keras-applications=1.0.8=pypi_0
keras-preprocessing=1.1.2=pypi_0
libedit=3.1.20210216=h27cfd23_1
libffi=3.2.1=hf484d3e_1007
libgcc-ng=9.1.0=hdf63c60_0
libstdcxx-ng=9.1.0=hdf63c60_0
markdown=3.2.2=pypi_0
mkl=2018.0.3=1
mkl-service=1.1.2=py35h90e4bf4_5
ncurses=6.2=he6710b0_1
numpy=1.18.5=pypi_0
openssl=1.0.2u=h7b6447c_0
pillow=7.2.0=pypi_0
pip=10.0.1=py35_0
protobuf=3.15.6=pypi_0
python=3.5.6=hc3d631a_0
pyyaml=5.3.1=pypi_0
readline=7.0=h7b6447c_5
scipy=1.4.1=pypi_0
setuptools=50.3.2=pypi_0
six=1.15.0=pypi_0
sqlite=3.33.0=h62c20be_0
tensorboard=1.14.0=pypi_0
tensorflow=1.14.0=pypi_0
tensorflow-estimator=1.14.0=pypi_0
termcolor=1.1.0=pypi_0
theano=1.0.4=pypi_0
tk=8.6.10=hbc83047_0
werkzeug=1.0.1=pypi_0
wheel=0.36.2=pyhd3eb1b0_0
wrapt=1.12.1=pypi_0
xz=5.2.5=h7b6447c_0
zipp=1.2.0=pypi_0